Refactor Params
